+++ libc/utils/testutils/CMakeLists.txt
@@ -5,8 +5,6 @@ endif()
 
 add_library(
   libc_test_utils
-  RandUtils.cpp
-  RandUtils.h
   StreamWrapper.cpp
   StreamWrapper.h
   ${EFFile}
